# openapi.yaml
openapi: 3.0.3
info:
  title: Fulfillment Order Management API
  description: API for managing the lifecycle of fulfillment orders.
  version: 1.0.0
servers:
  - url: https://api.example.com/fulfillment/order_management/v1
paths:
  /fulfillment_orders:
    post:
      summary: Create a new Fulfillment Order
      description: |
        Submits a new fulfillment order to the system. 
        Clients must provide an `Idempotency-Key` header so the service can safely retry the request without creating duplicate orders. 
        The payload's `seller_fulfillment_order_id` still needs to be unique; if an order with the same ID already exists, a 409 Conflict will be returned.
      operationId: create_fulfillment_order
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: A unique key generated by the client to guarantee that retries of this request are processed only once.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFulfillmentOrderRequest'
      responses:
        '202':
          description: Accepted. The fulfillment order has been successfully received and is being processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FulfillmentOrder'
        '400':
          description: Bad Request. The request body is invalid or missing required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '409':
          description: Conflict. A fulfillment order with the provided `seller_fulfillment_order_id` already exists.
  /fulfillment_orders/{order_id}:
    get:
      summary: Get Fulfillment Order by ID
      description: Retrieves the current state and details of a specific fulfillment order using its system-generated unique ID.
      operationId: get_fulfillment_order_by_id
      parameters:
        - name: order_id
          in: path
          required: true
          description: The unique system-generated identifier for the fulfillment order.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FulfillmentOrder'
        '404':
          description: Not Found. No fulfillment order found with the given ID.
  /fulfillment_orders/{order_id}/cancel:
    post:
      summary: Cancel a Fulfillment Order
      description: Requests the cancellation of a fulfillment order. Cancellation is only possible if the order has not yet been shipped.
      operationId: cancel_fulfillment_order
      parameters:
        - name: order_id
          in: path
          required: true
          description: The unique system-generated identifier for the fulfillment order.
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Accepted. The cancellation request has been accepted and is being processed.
        '400':
          description: Bad Request. The order is not in a state that allows cancellation (e.g., already shipped).
        '404':
          description: Not Found. No fulfillment order found with the given ID.

components:
  schemas:
    CreateFulfillmentOrderRequest:
      type: object
      required:
        - seller_fulfillment_order_id
        - displayable_order_id
        - displayable_order_date
        - shipping_speed_category
        - destination_address
        - items
      properties:
        seller_fulfillment_order_id:
          type: string
          description: A unique identifier for the order, provided by the seller. Used for idempotency.
          maxLength: 255
          example: "seller-abc-123"
        displayable_order_id:
          type: string
          description: An identifier for the order that is shown to the recipient on packing slips.
          maxLength: 255
          example: "cust-po-456"
        displayable_order_date:
          type: string
          format: date-time
          description: The date of the order to be displayed on packing slips. Cannot be in the future.
        displayable_order_comment:
          type: string
          description: A comment for the order that is shown to the recipient.
          maxLength: 500
          example: "Thank you for your order!"
        shipping_speed_category:
          type: string
          description: The requested shipping speed.
          enum:
            - STANDARD
            - EXPEDITED
            - PRIORITY
            - SAME_DAY
            - NEXT_DAY
            - SCHEDULED
          example: "STANDARD"
        destination_address:
          $ref: '#/components/schemas/Address'
        items:
          type: array
          description: List of items to fulfill. Must contain at least 1 item and no more than 100 items.
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/OrderItem'

    FulfillmentOrder:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
          description: The unique system-generated identifier for the fulfillment order.
        seller_fulfillment_order_id:
          type: string
        displayable_order_id:
          type: string
        displayable_order_date:
          type: string
          format: date-time
        displayable_order_comment:
          type: string
        shipping_speed_category:
          type: string
        destination_address:
          $ref: '#/components/schemas/Address'
        status:
          type: string
          description: The current status of the fulfillment order.
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        received_date:
          type: string
          format: date-time
          description: The timestamp when the order was received by the system.

    OrderItem:
      type: object
      required:
        - seller_sku
        - seller_fulfillment_order_item_id
        - quantity
      properties:
        seller_sku:
          type: string
          description: The seller's unique identifier for the item.
          maxLength: 255
          example: "item-xyz-789"
        seller_fulfillment_order_item_id:
          type: string
          description: The seller's unique identifier for the line item.
          maxLength: 255
          example: "line-item-1"
        quantity:
          type: integer
          format: int32
          description: Quantity to fulfill. Must be between 1 and 10,000.
          minimum: 1
          maximum: 10000
        gift_message:
          type: string
          description: Optional gift message to include with the item.
          maxLength: 500
        displayable_comment:
          type: string
          description: Optional comment to display on packing slip.
          maxLength: 500

    Address:
      type: object
      required:
        - name
        - address_line_1
        - city
        - state_or_region
        - postal_code
        - country_code
      properties:
        name:
          type: string
          description: Recipient name.
          maxLength: 255
          example: "John Doe"
        address_line_1:
          type: string
          description: First line of the address.
          maxLength: 255
          example: "123 Main Street"
        address_line_2:
          type: string
          description: Second line of the address (optional).
          maxLength: 255
          example: "Apt 4B"
        city:
          type: string
          description: City name.
          maxLength: 100
          example: "New York"
        state_or_region:
          type: string
          description: State, province, or region.
          maxLength: 100
          example: "NY"
        postal_code:
          type: string
          description: Postal or ZIP code. Must contain only alphanumeric characters, spaces, and hyphens.
          maxLength: 20
          pattern: '^[A-Za-z0-9\s-]+$'
          example: "10001"
        country_code:
          type: string
          description: Two-letter ISO 3166-1 alpha-2 country code (uppercase).
          minLength: 2
          maxLength: 2
          pattern: '^[A-Z]{2}$'
          example: "US"

    ValidationErrorResponse:
      type: object
      description: Error response returned when request validation fails.
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred.
        status:
          type: integer
          description: HTTP status code.
          example: 400
        error:
          type: string
          description: Error type.
          example: "Validation Failed"
        message:
          type: string
          description: Human-readable error message.
          example: "Request validation failed. Please check the errors for details."
        path:
          type: string
          description: Request path that caused the error.
          example: "/fulfillment_orders"
        fieldErrors:
          type: object
          description: Map of field names to their validation error messages.
          additionalProperties:
            type: array
            items:
              type: string
          example:
            seller_fulfillment_order_id:
              - "Seller fulfillment order ID is required"
            items:
              - "Order must have at least one item"
            destination_address.postal_code:
              - "Postal code contains invalid characters"